dating-app-backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# MongoDB
motor==3.3.2
pymongo==4.6.0

# Authentication
pyjwt==2.8.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0

# Environment
python-dotenv==1.0.0

# Image Storage
cloudinary==1.38.0

# Validation & Sanitization
pydantic==2.5.0
pydantic-settings==2.1.0
bleach==6.1.0

# Rate Limiting
slowapi==0.1.7

# WebSocket
websockets==12.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
